Corriger le type de l'échelle de la minicarte
La division d'une hauteur en Float par une portée en Double produisait un Double,
refusé par la projection et la barre d'échelle qui attendent des Float.

+++ app/src/main/kotlin/io/github/jmallus/guidage/ui/MapRenderer.kt
@@ -47,7 +47,8 @@ object MapRenderer {
 
         val riderX = area.centerX()
         val riderY = area.top + area.height() * AHEAD_FRACTION
-        val metersToPixels = (area.height() * AHEAD_FRACTION) / model.rangeMeters.coerceAtLeast(1.0)
+        val metersToPixels =
+            ((area.height() * AHEAD_FRACTION) / model.rangeMeters.coerceAtLeast(1.0)).toFloat()
         val heading = model.heading ?: 0.0
 
         val projection = Projection(origin, heading, riderX, riderY, metersToPixels)
